Rename the show-all-schools controller to schools-list
It is about to own searching and the no-results state as well as the
collapse, so the name no longer describes it.
