+++ app/javascript/publish/application.js
@@ -7,7 +7,7 @@ import FilterToggle from './filters'
 import { Application } from '@hotwired/stimulus'
 import InputPreviewController from './courses/input_preview_controller'
 import SelectAllCheckboxesController from './controllers/select_all_checkboxes_controller'
-import ShowAllSchoolsController from './controllers/show_all_schools_controller'
+import SchoolsListController from './controllers/schools_list_controller'
 import CopyLinkController from './controllers/copy_link_controller'
 import RemoteAutocompleteController from '../shared/remote_autocomplete_controller'
 
@@ -20,6 +20,6 @@ FilterToggle.init()
 window.Stimulus = Application.start()
 Stimulus.register('input-preview', InputPreviewController)
 Stimulus.register('select-all-checkboxes', SelectAllCheckboxesController)
-Stimulus.register('show-all-schools', ShowAllSchoolsController)
+Stimulus.register('schools-list', SchoolsListController)
 Stimulus.register('copy-link', CopyLinkController)
 Stimulus.register('remote-autocomplete', RemoteAutocompleteController)
