+++ app/javascript/publish/controllers/schools_list_controller.js
