+++ app/views/publish/course_wizards/steps/_schools.html.erb
@@ -27,7 +27,7 @@
 
 <%= form.govuk_check_boxes_fieldset :school_uuids,
       legend: { hidden: true, text: heading },
-      form_group: { data: { controller: "show-all-schools", show_all_schools_visible_value: current_step.schools_collapse_threshold } } do %>
+      form_group: { data: { controller: "schools-list", schools_list_visible_value: current_step.schools_collapse_threshold } } do %>
   <% current_step.schools.each_with_index do |school, index| %>
     <%= form.govuk_check_box :school_uuids,
           school.uuid,
@@ -35,15 +35,15 @@
           hint: { text: school.decorate.full_address, class: "govuk-!-font-size-16" },
           include_hidden: false,
           link_errors: index.zero?,
-          data: { show_all_schools_target: "school" } %>
+          data: { schools_list_target: "school" } %>
   <% end %>
 
   <% if current_step.collapse_schools? %>
     <button type="button"
             class="app-button-link govuk-!-font-size-19 govuk-!-margin-top-4"
             hidden
-            data-show-all-schools-target="showAll"
-            data-action="show-all-schools#showAll">
+            data-schools-list-target="showAll"
+            data-action="schools-list#showAll">
       <%= t("publish.courses.schools.show_all") %>
     </button>
   <% end %>
